requirements.txt
# OpenAI API
openai==1.3.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Web Interface
streamlit==1.28.0

# Environment Variables
python-dotenv==1.0.0

# Utilities
requests==2.31.0